refactor(lwm): consolidate notifications-prompt triggers onto notifyFlowCompleted
Migrate every remaining after-action trigger call site (onboarding, send,
receive, swap, stake, add-favorite-coin, WalletSync, Settings, and the
NotificationsOptIn screen) from the legacy useNotifications hook chain onto
the single notifyFlowCompleted / useNotificationsContext production path,
then delete the now-dead useNotifications/useNotificationsPrompt/
useNotificationsDrawer hooks and shouldPromptOptInDrawerAfterAction.
