+++ apps/ledger-live-mobile/src/mvvm/features/Market/__integrations__/shared.tsx
@@ -4,6 +4,7 @@ import type { MarketNavigatorStackParamList } from "LLM/features/Market/Navigato
 import WalletTabNavigatorScrollManager from "~/components/WalletTab/WalletTabNavigatorScrollManager";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { ScreenName } from "~/const";
+import { NotificationsPromptProvider } from "LLM/features/NotificationsPrompt";
 import MarketList from "../screens/MarketList";
 import MarketDetail from "../screens/MarketDetail";
 import MarketCurrencySelect from "../screens/MarketCurrencySelect";
@@ -12,15 +13,17 @@ const StackWalletTab = createNativeStackNavigator<MarketNavigatorStackParamList>
 
 export function MarketPages() {
   return (
-    <WalletTabNavigatorScrollManager>
-      <StackWalletTab.Navigator initialRouteName={ScreenName.MarketList}>
-        <StackWalletTab.Screen name={ScreenName.MarketList} component={MarketList} />
-        <StackWalletTab.Screen name={ScreenName.MarketDetail} component={MarketDetail} />
-        <StackWalletTab.Screen
-          name={ScreenName.MarketCurrencySelect}
-          component={MarketCurrencySelect}
-        />
-      </StackWalletTab.Navigator>
-    </WalletTabNavigatorScrollManager>
+    <NotificationsPromptProvider>
+      <WalletTabNavigatorScrollManager>
+        <StackWalletTab.Navigator initialRouteName={ScreenName.MarketList}>
+          <StackWalletTab.Screen name={ScreenName.MarketList} component={MarketList} />
+          <StackWalletTab.Screen name={ScreenName.MarketDetail} component={MarketDetail} />
+          <StackWalletTab.Screen
+            name={ScreenName.MarketCurrencySelect}
+            component={MarketCurrencySelect}
+          />
+        </StackWalletTab.Navigator>
+      </WalletTabNavigatorScrollManager>
+    </NotificationsPromptProvider>
   );
 }
