+++ apps/ledger-live-mobile/src/mvvm/features/NotificationsPrompt/index.ts
@@ -1,19 +1,15 @@
-export { useNotifications } from "./hooks/useNotifications";
 export { NotificationsPromptWrapper } from "./new/NotificationsPromptWrapper";
 export { NotificationsPromptProvider } from "./new/NotificationsPromptProvider";
 export { useNotificationsContext } from "./new/NotificationsPromptProvider";
 export { useNotificationsData } from "./hooks/useNotificationsData";
-export { useNotificationsPrompt } from "./hooks/useNotificationsPrompt";
-export { useNotificationsDrawer } from "./hooks/useNotificationsDrawer";
 export {
+  AFTER_ACTION_SOURCE_TO_EVENT_KEY,
   INACTIVITY_DRAWER_DELAY_MS,
-  checkIsInactive,
   evaluateAfterActionTrigger,
   evaluateInactivityTrigger,
   getNextRepromptDelay,
   canPromptTransactionsAlertsForAction,
   getNotificationPromptTarget,
-  shouldPromptOptInDrawerAfterAction,
 } from "./utils/notificationsPromptEngine";
 export type {
   DataOfUser,
