+++ apps/ledger-live-mobile/src/screens/Settings/Notifications/index.tsx
@@ -10,7 +10,8 @@ import { track, TrackScreen, trackWithRoute, updateIdentify } from "~/analytics"
 import { notificationsSelector } from "~/reducers/settings";
 import { setNotifications } from "~/actions/settings";
 import type { State } from "~/reducers/types";
-import { useNotifications } from "LLM/features/NotificationsPrompt";
+import { useNotificationsData } from "LLM/features/NotificationsPrompt";
+import { useNotificationsPermission } from "LLM/hooks/useNotificationsPermission";
 import { updateUserPreferences } from "~/notifications/braze";
 import { useFeature } from "@features/platform-feature-flags";
 import { AuthorizationStatus } from "@react-native-firebase/messaging";
@@ -34,7 +35,8 @@ type NotificationRowProps = {
 function NotificationSettingsRow({ disabled, notificationKey, label }: NotificationRowProps) {
   const dispatch = useDispatch();
   const notifications = useSelector(notificationsSelector);
-  const { markUserAsOptIn, permissionStatus, markUserAsOptOut } = useNotifications();
+  const { markUserAsOptIn, markUserAsOptOut } = useNotificationsData();
+  const { permissionStatus } = useNotificationsPermission();
 
   const { t } = useTranslation();
 
@@ -96,7 +98,7 @@ function NotificationSettingsRow({ disabled, notificationKey, label }: Notificat
 function NotificationsSettings() {
   const { t } = useTranslation();
   const notifications = useSelector(notificationsSelector);
-  const { permissionStatus, requestPushNotificationsPermission } = useNotifications();
+  const { permissionStatus, requestPushNotificationsPermission } = useNotificationsPermission();
 
   const featureBrazePushNotifications = useFeature("brazePushNotifications");
   const hiddenNotificationCategories = useMemo(() => {
